S3_trigger_funtion/lambda_function/requirements.txt
boto3==1.18.48
botocore==1.21.48
openpyxl==3.0.9
pip==21.2.4
pytz==2021.1
urllib3==1.26.7
Pillow == 8.3.2
